+++ .agents/skills/ml-property-predict-scd/templates/finetune_config.template.yaml
@@ -1,5 +1,5 @@
 # Full-model finetuning only.
-# For a frozen-backbone linear probe, use templates/train_linear_probe_head.py instead.
+# For lightweight frozen-backbone head training, use templates/train_lightweight_head.py instead.
 log_dir: experiments/
 load_model: null
 load_hf: null
@@ -78,28 +78,3 @@ rep_min_atoms: 0
 random_rotate: false
 torsion_angle_std: 0.0
 max_bonds_rotated: 0
-p_invert: 0.0
-
-# For downstream scalar prediction from a pretrained SCD checkpoint,
-# resetting the head is usually a sensible default.
-reset_head: true
-reset_embeddings: false
-reset_norms: null
-
-droppath_warmup: true
-init_droppath: 0.15
-final_droppath: 0.1
-beta1: 0.995
-
-momentum_update: false
-model_momentum: 0.996
-val_ema_model: false
-
-# Force-aware training
-# derivative: true
-# direct_force_pred: false
-
-# Optional OOM protection
-# max_nodes_per_batch: 4000
-# batch_clipper_cache_size: 1000
-# allow_test_clipping: false
